// MCPPortal/src/app/_modules/enter/app/core/interfaces/object.interface.ts
export interface Object {
  mainObject: string;
  object: string;
  subObject: string;
  locatedAtPartyAddress: string;
  party: string;
  damageLocation: string;
  coverage: string;
  policyNumber: string;
  deductibleExcess: string;
  insuredAmount: string;
  policyConditions: string;
  deedOfAssignment: string;
  startDateKnown: boolean;
  policyDuration: string;
  guaranteeAgainstUnderinsurance: boolean;
  includingExcludingVAT: string;
}